replace individual listeners for each command with a single listener in Commands
