+++ src/main/java/com/robothand/highqualitybot/command/ClearCommand.java
@@ -33,7 +33,7 @@ public class ClearCommand extends Command {
     }
 
     @Override
-    public void onCommand(MessageReceivedEvent event, String[] args) {
+    public void execute(MessageReceivedEvent event, String[] args) {
         GuildMusicPlayer musicPlayer;
         Guild guild = event.getGuild();
         MessageChannel channel = event.getChannel();
