+++ src/main/java/com/robothand/highqualitybot/command/HelpCommand.java
@@ -31,7 +31,7 @@ public class HelpCommand extends Command {
     }
 
     @Override
-    public void onCommand(MessageReceivedEvent event, String[] args) {
+    public void execute(MessageReceivedEvent event, String[] args) {
         MessageChannel channel = event.getChannel();
         String message;
