+++ src/main/java/com/robothand/highqualitybot/command/PingCommand.java
@@ -28,7 +28,7 @@ public class PingCommand extends Command {
     }
 
     @Override
-    public void onCommand(MessageReceivedEvent event, String[] args) {
+    public void execute(MessageReceivedEvent event, String[] args) {
         MessageChannel channel = event.getChannel();
 
         channel.sendMessage("Pong!").queue();
