+++ src/main/java/com/robothand/highqualitybot/command/RepeatCommand.java
@@ -30,7 +30,7 @@ public class RepeatCommand extends Command {
     }
 
     @Override
-    public void onCommand(MessageReceivedEvent event, String[] args) {
+    public void execute(MessageReceivedEvent event, String[] args) {
         GuildMusicPlayer musicPlayer;
         Guild guild = event.getGuild();
         MessageChannel channel = event.getChannel();
