+++ src/main/java/com/robothand/highqualitybot/command/SkipCommand.java
@@ -28,7 +28,7 @@ public class SkipCommand extends Command {
     }
 
     @Override
-    public void onCommand(MessageReceivedEvent event, String[] args) {
+    public void execute(MessageReceivedEvent event, String[] args) {
         GuildMusicPlayer musicPlayer;
         Guild guild = event.getGuild();
         MessageChannel channel = event.getChannel();
