+++ src/main/java/com/robothand/highqualitybot/command/StatusCommand.java
@@ -26,7 +26,7 @@ public class StatusCommand extends Command {
     }
 
     @Override
-    public void onCommand(MessageReceivedEvent event, String[] args) {
+    public void execute(MessageReceivedEvent event, String[] args) {
         MessageChannel channel = event.getChannel();
         StringBuilder messageBuilder = new StringBuilder();
